refactor: unify product card component
