+++ app/components/featured-products-grid.module.css
@@ -5,116 +5,6 @@
   gap: 12px;
 }
 
-.card {
-  border-radius: 10px;
-  border: 1px solid #dfe5ec;
-  background: transparent;
-  overflow: hidden;
-  cursor: pointer;
-}
-
-.card:focus-visible {
-  outline: 2px solid #5adf26;
-  outline-offset: 2px;
-  border-color: #b9ebc6;
-}
-
-.imageWrap {
-  position: relative;
-  width: 100%;
-  aspect-ratio: 4 / 5;
-  background: transparent;
-}
-
-.image {
-  object-fit: cover;
-  object-position: center;
-}
-
-.cardBody {
-  background: #fff;
-  padding: 11px;
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-  min-height: 118px;
-}
-
-.cardBody h3 {
-  margin: 0 0 1px;
-  font-family: var(--font-display), "Trebuchet MS", sans-serif;
-  font-size: 16px;
-  line-height: 1.1;
-  letter-spacing: -0.02em;
-  min-height: calc(1.1em * 2);
-  display: -webkit-box;
-  -webkit-line-clamp: 2;
-  -webkit-box-orient: vertical;
-  overflow: hidden;
-}
-
-.cardBottom {
-  display: grid;
-  gap: 6px;
-  margin-top: auto;
-}
-
-.priceStack {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  min-height: 0;
-}
-
-.previousPrice {
-  color: #7c8799;
-  font-size: 13px;
-  line-height: 1;
-  text-decoration: line-through;
-}
-
-.previousPriceEmpty {
-  display: none;
-}
-
-.cardBottom strong {
-  display: block;
-  color: #21cb16;
-  font-size: clamp(18px, 1.45vw, 23px);
-  letter-spacing: -0.02em;
-  line-height: 1.05;
-}
-
-.cardBottom button {
-  width: 100%;
-  min-height: 34px;
-  border-radius: 10px;
-  border: 1px solid #a5e7b6;
-  background: #dff6e5;
-  color: #1dbb17;
-  font-weight: 700;
-  font-size: 16px;
-  font-family: var(--font-display), "Trebuchet MS", sans-serif;
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  gap: 6px;
-  padding: 0 10px;
-  white-space: nowrap;
-  cursor: pointer;
-  transition: border-color 0.14s ease;
-}
-
-.cardBottom button:hover {
-  border-color: #5adf26;
-}
-
-.cardBottom button svg {
-  width: 16px;
-  height: 16px;
-  flex-shrink: 0;
-}
-
 @media (max-width: 1100px) {
   .grid {
     grid-template-columns: repeat(2, minmax(0, 1fr));
@@ -126,30 +16,6 @@
   .grid {
     grid-template-columns: repeat(2, minmax(0, 1fr));
     gap: 8px;
-  }
-
-  .cardBottom button {
-    font-size: 14px;
-    gap: 5px;
-    padding-inline: 8px;
-  }
-
-  .cardBody {
-    min-height: 106px;
-  }
-}
-
-@media (max-width: 520px) {
-  .cardBottom button {
-    font-size: 12px;
-    min-height: 32px;
-    gap: 4px;
-    letter-spacing: 0;
-  }
-
-  .cardBottom button svg {
-    width: 14px;
-    height: 14px;
   }
 }
