+++ app/components/product-card.module.css
@@ -1,0 +1,211 @@
+.card {
+  border-radius: 10px;
+  border: 1px solid #dfe5ec;
+  background: transparent;
+  overflow: hidden;
+  cursor: pointer;
+  transition:
+    transform 0.18s ease,
+    box-shadow 0.18s ease,
+    border-color 0.18s ease;
+}
+
+.cardReveal {
+  opacity: 0;
+  transform: translateY(18px);
+  will-change: opacity, transform;
+  transition:
+    box-shadow 0.18s ease,
+    border-color 0.18s ease;
+}
+
+.cardVisible {
+  animation: cardFadeUp 0.48s cubic-bezier(0.16, 1, 0.3, 1) both;
+  animation-delay: var(--reveal-delay, 0ms);
+}
+
+@keyframes cardFadeUp {
+  from {
+    opacity: 0;
+    transform: translateY(18px);
+  }
+  to {
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
+
+.card:hover {
+  transform: translateY(-2px);
+  box-shadow: 0 16px 30px rgba(22, 35, 56, 0.09);
+}
+
+.card:focus-visible {
+  outline: 2px solid #5adf26;
+  outline-offset: 2px;
+  border-color: #b9ebc6;
+}
+
+.imageWrap {
+  position: relative;
+  width: 100%;
+  aspect-ratio: 4 / 5;
+  background: transparent;
+}
+
+.image {
+  object-fit: cover;
+  object-position: center;
+  transition: opacity 260ms ease;
+}
+
+.imagePrimary,
+.imageSecondary {
+  opacity: 1;
+}
+
+.imageSecondary {
+  opacity: 0;
+}
+
+.imageHidden {
+  opacity: 0;
+}
+
+.imageVisible {
+  opacity: 1;
+}
+
+.cardBody {
+  background: #fff;
+  padding: 11px;
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+  min-height: 118px;
+}
+
+.cardBody h2 {
+  margin: 0 0 1px;
+  font-family: var(--font-display), "Trebuchet MS", sans-serif;
+  font-size: 16px;
+  line-height: 1.1;
+  letter-spacing: -0.02em;
+  min-height: calc(1.1em * 2);
+  display: -webkit-box;
+  -webkit-line-clamp: 2;
+  -webkit-box-orient: vertical;
+  overflow: hidden;
+}
+
+.cardBottom {
+  display: grid;
+  gap: 6px;
+  margin-top: auto;
+}
+
+.priceStack {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  min-height: 0;
+}
+
+.previousPrice {
+  color: #7c8799;
+  font-size: 13px;
+  line-height: 1;
+  text-decoration: line-through;
+}
+
+.previousPriceEmpty {
+  display: none;
+}
+
+.cardBottom strong {
+  display: block;
+  color: #21cb16;
+  font-size: clamp(18px, 1.45vw, 23px);
+  letter-spacing: -0.02em;
+  line-height: 1.05;
+}
+
+.cardBottom button {
+  width: 100%;
+  min-height: 34px;
+  padding: 7px 10px;
+  border-radius: 9px;
+  border: 1px solid #c8f2d0;
+  background: #ecfff0;
+  color: #1ec015;
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  gap: 6px;
+  font-family: var(--font-display), "Trebuchet MS", sans-serif;
+  font-size: 12px;
+  font-weight: 700;
+  letter-spacing: 0.02em;
+  cursor: pointer;
+}
+
+.cardBottom button svg {
+  width: 14px;
+  height: 14px;
+  stroke-width: 2.2;
+}
+
+@media (max-width: 1100px) {
+  .cardBody h2 {
+    font-size: 15px;
+  }
+
+  .cardBottom strong {
+    font-size: 19px;
+  }
+}
+
+@media (max-width: 760px) {
+  .cardBody {
+    padding: 9px;
+    min-height: 106px;
+  }
+
+  .cardBody h2 {
+    font-size: 14px;
+  }
+
+  .cardBottom strong {
+    font-size: 17px;
+  }
+
+  .cardBottom button {
+    min-height: 32px;
+    font-size: 11px;
+    letter-spacing: 0;
+  }
+}
+
+@media (max-width: 520px) {
+  .imageWrap {
+    aspect-ratio: 3 / 4;
+  }
+
+  .cardBody {
+    min-height: 0;
+  }
+}
+
+@media (prefers-reduced-motion: reduce) {
+  .image {
+    transition: none;
+  }
+
+  .cardReveal {
+    opacity: 1;
+    transform: none;
+    transition:
+      box-shadow 0.18s ease,
+      border-color 0.18s ease;
+  }
+}
